revert: Revert packet name change from 48e2cf6b42da69d11913ac26316f78cab99ad9ee
