+++ pumpkin-protocol/src/client/play/mod.rs
@@ -23,6 +23,7 @@ mod disguised_chat_message;
 mod display_objective;
 mod entity_animation;
 mod entity_metadata;
+mod entity_position_sync;
 mod entity_sound_effect;
 mod entity_status;
 mod entity_velocity;
@@ -69,7 +70,6 @@ mod spawn_entity;
 mod stop_sound;
 mod store_cookie;
 mod subtitle;
-mod synchronize_vehicle_position;
 mod system_chat_message;
 mod take_item;
 mod teleport_entity;
@@ -108,6 +108,7 @@ pub use disguised_chat_message::*;
 pub use display_objective::*;
 pub use entity_animation::*;
 pub use entity_metadata::*;
+pub use entity_position_sync::*;
 pub use entity_sound_effect::*;
 pub use entity_status::*;
 pub use entity_velocity::*;
@@ -154,7 +155,6 @@ pub use spawn_entity::*;
 pub use stop_sound::*;
 pub use store_cookie::*;
 pub use subtitle::*;
-pub use synchronize_vehicle_position::*;
 pub use system_chat_message::*;
 pub use take_item::*;
 pub use teleport_entity::*;
